requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
Flask==3.0.3
flask-paginate==2024.3.28
flask-restx==1.3.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.41
langchain-mongodb==0.1.3
langchain-openai==0.1.2
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.43
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.16.2
orjson==3.10.0
packaging==23.2
priority==2.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
Quart==0.19.5
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
sniffio==1.3.1
SQLAlchemy==2.0.29
taskgroup==0.0.0a4
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.2
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
